+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Change Log _InterSystems SAP Service_
 
 ---
+## [v0.1](https://github.com/intersystems-dach/intersystems-sap-service/tree/0.1) (2024-03-11)
 
-## [v1.0.0](https://github.com/intersystems-dach/intersystems-sap-service/tree/1.0.0) (2023-4-3)
+- Adds compatibility to InterSystems IRIS version 2022.1 and newer
+- Removes intersystems-utils and intersystems-jdbr java library from jar. Intersystems-utils must be added to class path manually.
+- Use v0.0.6 for InterSystems IRIS versions older than 2022.1
+
+## [v0.0.6](https://github.com/intersystems-dach/intersystems-sap-service/tree/0.0.6) (2023-4-3)
 
 -   _Initial release_